MINOR: refactor configuration of RestService
This will make it easier for RestService to be shared by other client
implementations.
